+++ apps/backend/tests/testutil/gateway/store_setup.go
@@ -38,9 +38,9 @@ func applyGatewayCompanyState(t *testing.T, cfg config.Config, st store.Store, o
 	t.Helper()
 	ctx := testutil.CtxForCompany(opts.CompanyID)
 
-	walletPoint := 100000.0
+	walletPoint := int64(100000)
 	if opts.WalletBalancePoint != nil {
-		walletPoint = *opts.WalletBalancePoint
+		walletPoint = int64(*opts.WalletBalancePoint)
 	}
 	if err := st.Company().SetWalletRemain(ctx, opts.CompanyID, walletPoint, nil); err != nil {
 		t.Fatal(err)
